Use name, not magic ID, for DIBS metaclass init
The parent to be used when creating the DIBS payment metaclass was fetched
based on a hardcoded ID. The DIBS class is always created, but possibly under the
wrong parent, if this hardcoded ID points to the wrong class. This varies from
environment to environment. This commits fixes this, by fetching the
parent by name instead of ID.